app/requirements.txt
Click==7.0
dependency-injector==3.14.5
Flask==1.0.2
gunicorn==19.7.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.2
mysql-connector-python==8.0.16
peewee==3.9.5
protobuf==3.7.1
PyMySQL==0.9.3
python-dotenv==0.10.2
six==1.12.0
Werkzeug==0.15.2
pytest==4.5.0